moves message reception logic to app component
