+++ src/app/views/App.tsx
@@ -10,7 +10,6 @@ import { connect } from 'react-redux';
 import { bindActionCreators, Dispatch } from 'redux';
 
 import { loadGETheme } from '../../themes';
-import { dark } from '../../themes/dark';
 import { Mode } from '../../types/action';
 import { IInitMessage, IThemeChangedMessage } from '../../types/query-runner';
 import { clearQueryError } from '../services/actions/error-action-creator';
@@ -22,6 +21,7 @@ import { Authentication } from './authentication';
 import { classNames } from './classnames';
 import { QueryResponse } from './query-response';
 import { QueryRunner } from './query-runner';
+import { parse } from './query-runner/util/iframe-message-parser';
 import { Sidebar } from './sidebar/Sidebar';
 
 interface IAppProps {
@@ -50,17 +50,47 @@ class App extends Component<IAppProps, IAppState> {
   }
 
   public componentDidMount = () => {
-    // TODO: delete these calls and and make them where relevant.
-    this.props.actions!.clearQueryError();
-    this.props.actions!.runQuery();
-    this.props.actions!.addRequestHeader();
-    this.props.actions!.setSampleQuery();
+    const { actions } = this.props;
+    const whiteListedDomains = [
+      'https://docs.microsoft.com',
+      'https://review.docs.microsoft.com',
+      'https://ppe.docs.microsoft.com',
+      'https://docs.azure.cn'
+    ];
+
+    // Notify host document that GE is ready to receive messages
+    const hostOrigin = new URLSearchParams(location.search).get('host-origin');
+    const originIsWhitelisted =
+      hostOrigin && whiteListedDomains.indexOf(hostOrigin) !== -1;
+
+    if (hostOrigin && originIsWhitelisted) {
+      window.parent.postMessage({ type: 'ready' }, hostOrigin);
+    }
 
     // Listens for messages from host document
     window.addEventListener('message', this.receiveMessage, false);
-    // tslint:disable-next-line:no-console
-    console.log('adding listener');
-  }
+
+    const urlParams = new URLSearchParams(window.location.search);
+    const base64Token = urlParams.getAll('query')[0];
+
+    if (!base64Token) {
+      return;
+    }
+
+    const data = JSON.parse(atob(base64Token));
+    const { sampleVerb, sampleHeaders, sampleUrl, sampleBody } = data;
+
+    const query = {
+      sampleUrl,
+      sampleBody,
+      sampleHeaders,
+      selectedVerb: sampleVerb
+    };
+
+    if (actions) {
+      actions.setSampleQuery(query);
+    }
+  };
 
   public componentWillUnmount(): void {
     window.removeEventListener('message', this.receiveMessage);
@@ -78,11 +108,53 @@ class App extends Component<IAppProps, IAppState> {
     const msgEvent: IThemeChangedMessage | IInitMessage = event.data;
 
     switch (msgEvent.type) {
+      case 'init':
+        this.handleInitMsg(msgEvent);
+        break;
       case 'theme-changed':
         this.handleThemeChangeMsg(msgEvent);
         break;
       default:
         return;
+    }
+  };
+
+  private handleInitMsg = (msg: IInitMessage) => {
+    const { actions } = this.props;
+    const { verb, headers, url, body }: any = parse(msg.code);
+
+    if (actions) {
+      actions.setSampleQuery({
+        sampleUrl: url,
+        selectedVerb: verb
+      });
+    }
+
+    // Sets selected verb in App Component
+    this.handleSelectVerb(verb);
+
+    /**
+     * We are delaying this by 1 second here so that we give Monaco's formatter time to initialize.
+     * If we don't put this delay, the body won't be formatted.
+     */
+    setTimeout(() => {
+      if (actions) {
+        actions.setSampleQuery({
+          sampleUrl: url,
+          selectedVerb: verb,
+          sampleBody: body
+        });
+      }
+    }, 1000);
+
+    if (actions) {
+      const requestHeadears = headers.map((header: any) => {
+        return {
+          name: Object.keys(header)[0],
+          value: Object.values(header)[0]
+        };
+      });
+      actions.addRequestHeader(requestHeadears);
     }
   };
